requirements.txt
BeautifulSoup4
pocketsphinx
pyaudio
vosk
piper
PyQt6
mysql-connector-python
